// packages/nosql/src/utils/types.ts
export interface INoSQLBase extends ICatalystGResponse {
	id: string;
	name: string;
	partition_key: INoSQLKeyColumn;
	sort_key?: INoSQLKeyColumn;
	type?: string; // find possible values for this
	status: 'CREATING' | 'ONLINE' | 'OFFLINE' | 'DELETED';
	metrics?: Record<string, unknown>;
}

export interface INoSQLKeyColumn {
	column_name: string;
	data_type: string | DataType;
}

export interface INoSQLTable extends INoSQLBase {
	ttl_enabled: boolean;
	api_access?: boolean;
	ttl_attribute?: string;
	additional_sort_keys: Array<INoSQLKeyColumn>;
	local_index?: Array<INoSQLIndex>;
	global_index: Array<INoSQLIndex>;
}

export interface INoSQLIndex extends INoSQLBase {
	id: string;
	name: string;
	projected_attributes: IProjectedAttributes;
}

export interface IProjectedAttributes {
	type: string | AttribType;
	include_columns?: Set<string>;
}